core: decompose ontology materializer workflows
